requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
requests==2.26.0

#rust based linter
ruff==0.0.284
pylint==2.15.3

#screenshot
pillow==10.0.1